requirements.txt
amqp==5.0.1
appnope==0.1.0
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.2
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.3
decorator==4.4.2
Django==2.2.16
django-debug-toolbar==3.1
django-extensions==3.0.8
et-xmlfile==1.0.1
Faker==4.14.1
flake8==3.8.3
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
gunicorn==20.0.4
html5lib==1.1
idna==2.10
iniconfig==1.1.1
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
kombu==5.0.2
lxml==4.5.2
mccabe==0.6.1
openpyxl==3.0.5
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-mock==3.3.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==5.0.4
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
XlsxWriter==1.3.7
django-redis==4.12.1
redis==3.5.3
djangorestframework==3.12.2
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0
pyjwt==1.7.1
